pyproject.toml
[project]
name = "univyx_backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Awolesi Oluwapemisin", email = "oluwapemisinawolesi@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "django-split-settings (>=1.3.2,<2.0.0)",
    "channels (>=4.2.0,<5.0.0)",
    "daphne (>=4.1.2,<5.0.0)",
    "channels-redis (>=4.2.1,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
